refactor: add domains folder
