+++ app/__tests__/vitest-setup.ts
@@ -15,8 +15,8 @@ vi.mock('next-intl', async (importOriginal) => {
   };
 });
 
-vi.mock('@/app/utils/get-validation-error', async (importOriginal) => {
-  const actual = await importOriginal<typeof import('@/app/utils/get-validation-error')>();
+vi.mock('@/app/domains/auth/get-validation-error', async (importOriginal) => {
+  const actual = await importOriginal<typeof import('@/app/domains/auth/get-validation-error')>();
   return {
     ...actual,
     getValidationError: vi.fn((tValidation, message) => (message ? tValidation(message) : '')),
