+++ app/_pages/SignIn.tsx
@@ -1,5 +1,5 @@
-import { SignInForm } from '@/app/components/SignInForm';
 import { PageWrapper } from '@/app/components/ui/PageWrapper';
+import { SignInForm } from '@/app/domains/auth/SignInForm';
 
 export const SignInPage = () => (
   <PageWrapper>
