+++ app/_pages/SignUp.tsx
@@ -1,5 +1,5 @@
-import { SignUpForm } from '@/app/components/SignUpForm';
 import { PageWrapper } from '@/app/components/ui/PageWrapper';
+import { SignUpForm } from '@/app/domains/auth/SignUpForm';
 
 export const SignUpPage = () => (
   <PageWrapper>
